span: Use publicly available member functions in operator==()
This is required if U != T, as the function is then no longer allowed to
access the other span's private members.

Also, comparing pointers to the object is rejected by the compiler if
U != T, so let's not do this. The performance penalty of this is
negligible, as the next to comparisons (for pointer and size) would
match, so no elements will be compared even if a span is compared to
itself.